[LEVEL3] Basic mmap syscall handling

+++ src/includes/defines.hh
@@ -121,8 +121,8 @@ public:
         void add_breakpoint(std::string, void* addr);
         ssize_t find_syscalls(void* addr);
         char is_from_us(void* addr) const;
-        int handle_bp(void* addr, bool print);
-        int exec_breakpoint(std::string region, void* addr, bool print);
+        long handle_bp(void* addr, bool print);
+        long exec_breakpoint(std::string region, void* addr, bool print);
         void print_bps() const;
         void reset_libs(void* link_map);
